Update patch set 1
Patch Set 1:

(13 comments)

I think it's a good idea. I've downloaded the patch and tried it out, and I have some problems with it:

1. The control string printed in the function 'end_colorize' is simply displayed on the terminal instead of changing the text color. (I'm using an Ubuntu 18.04 virtual image). However, if I change the ']' in the string to a '[', then it works.

2. The patch colorizes the entire output of the compiler to either red (errors), purple (warnings) or light blue (notes), with some notifications in the beginning.
I think it'd look better if only the strings 'error: ', 'warning: ', and 'note: ' were colorized (like in GCC). This way the output mostly has the default coloring (so it's easier to read), but the errors & warnings are easy to find, because of their color.
Alternatively the colorizing could be kept for the error/warning/note messages, if the errors are still too difficult to find with the above suggestion.

Patch-set: 1
CC: Gerrit User 1007689 <1007689@d5d70762-12d0-45a1-890d-524b12d3f735>